+++ dwds/test/extension_backend_test.dart
@@ -38,19 +38,19 @@ void main() {
         await ExtensionBackend.start(MockSocketHandler(), 'localhost');
   });
   test('returns success statusCode', () async {
-    var result = await http.get(Uri.parse(
+    final result = await http.get(Uri.parse(
         'http://localhost:${extensionBackend.port}/$authenticationPath'));
     expect(result.statusCode, 200);
   });
 
   test('returns expected authentication response', () async {
-    var result = await http.get(Uri.parse(
+    final result = await http.get(Uri.parse(
         'http://localhost:${extensionBackend.port}/$authenticationPath'));
     expect(result.body, authenticationResponse);
   });
 
   test('delegates to the underlying socket handler', () async {
-    var result = await http.get(
+    final result = await http.get(
         Uri.parse('http://localhost:${extensionBackend.port}/somedummypath'));
     expect(result.body, MockSocketHandler.mockResponse);
   });